Implement pushable objects. First one: Rock
Now when moving, if the destination tile is not empty, the board will
check if the object in the destination tile can be pushed by the moving
object. Each object defines the objects that it can and cannot be pushed
by.

Currently this allows the player to push rocks into empty space.
Anything else (e.g. player tries to push a wall, player tries to push a
rock into another object) will fail and no movement will occur.