Activate the player at a random position
The board now has a #random_blank_tile method, which will return a
random blank tile (duh). In order to keep this method simple, each
individual tile now needs to know its own x and y coordinates. This is
a bit of a bummer as I'd hoped to keep tiles pretty dumb, but an
individual tile's coordinates should never change so this hopefully
shouldn't come back to bite me.