Simplify Board#each_tile
Now that each tile knows its own coordinates, we don't have to work out
the coordinates within the #each_tile method and pass them to the
supplied block. Instead we can just pass the tile to the block, and the
block can ask the tile for its coordinates if needed.