Let rocks crush the player if they're in motion
A rock should not crush the player if the player moves into the tile
directly below the rock (ie. by crushing a patch of dirt that was
holding up the rock). However if a rock falls onto the player, then the
player should be crushed.

Objects now keep track of whether they're currently in motion or not
(based on whether their last chance to move was successful), and the
player checks if the object trying to crush it is currently in motion.
If not, it refuses to be crushed. Beautiful, isn't it?