Implement dropping grenades
Bluhhh big commit! Had to do some shuffling of board/tile functionality
to allow a grenade and a player to temporarily inhabit the same tile
when the player drops a grenade.

Pressing space now drops a grenade in the player's current tile, if the
player has any grenades available. The grenade then "explodes"
(currently this just prints "boom" to STDOUT) after 500 milliseconds,
and marks itself as expired. The tile's update loop will delete the
grenade when it detects that the grenade has expired.

Note that when I'm saying "grenade" above, I'm actually referring to a
LiveGrenade object. This objects uses the same sprite as a Grenade, but
has very different behaviour (cannot be collected, explodes and stuff)
so I've made it a completely different class.