Implement TargetScoreCalculator
Woo! We now calculate a target score for each game based on how many
points are actually available on the board.

Having this logic inside its own class should make it easy for us to
adjust the target percentage based on the selected difficulty level,
once that is a thing.