Only place Bombs on sturdy tiles
Our random placement was causing us to frequently place Bombs in places
where they would instantly move, causing them to blow up as soon as the
game started.

This change allows us to test if a tile is "sturdy" (ie. the contents of
the tile below are not slippable) when generating an object to place
there. We apply this logic to bombs. If the random object generator
generates a bomb for a tile that isn't sturdy, we put the bomb into a
backlog and we try it again on the next tile.

In order to make this possible, we now fill tiles from bottom-right to
top-left (instead of vice versa). It seems like this is the way the
original game generated tile contents, so that's cool!